вынос checkLengthText и getDescriptionsOfPhoto в модули
